Move code to more correct modules
- Move code that does not depend on boto3 and dynamDB outside of the
  dynamoDB specific module
