+++ Makefile
@@ -101,7 +101,7 @@ test-all: ## run tests on every Python version with tox
 	tox
 
 coverage: ## check code coverage quickly with the default Python
-	coverage run --source roachcase -m pytest
+	coverage run --source roachcase -m pytest -m 'not livedb'
 	coverage report -m
 	coverage html
 	$(BROWSER) htmlcov/index.html
